did more work on project viewer modal.
