+++ ui/src/components/Tabs/DashboardTab/TimelinesView/Timeline/ActionModal.js
@@ -520,6 +520,7 @@ export default function ActionModal(props) {
       <Modal
         closeOnDimmerClick={false}
         closeOnEscape={false}
+        closeIcon // close (makes x)
         className={"sticky"}
         onClose={() => {
           setOpen(false);
@@ -598,6 +599,7 @@ export default function ActionModal(props) {
       <Modal
         closeOnDimmerClick={false}
         closeOnEscape={false}
+        closeIcon // close (makes x)
         className={"sticky"}
         onClose={() => {
           setOpen(false);
