+++ ui/src/css/components/modal.css
@@ -299,7 +299,6 @@ body.dark-mode .ui.toggle.checkbox label:after {
   background-color: var(--text-primary) !important;
 }
 
-/* ===== MODAL SCROLLING FIX ===== */
 .ui.modal > .scrolling.content {
   max-height: calc(80vh - 10em);
   overflow-y: auto !important;
@@ -310,8 +309,6 @@ body.dark-mode .ui.modal > .scrolling.content {
   overflow-y: auto !important;
 }
 
-/* ===== FORCE CENTER ALL MODALS ===== */
-/* Higher specificity to override everything */
 body .ui.dimmer > .ui.modal,
 body.dark-mode .ui.dimmer > .ui.modal {
   position: fixed !important;
